feat: Simplify image middleware with express-sharp update
The express-sharp library has updated to v3.0.0, which includes a
[breaking
change](https://github.com/magento-research/express-sharp/pull/5) to the
URL scheme, and therefore the public API of the library.

This helps us simplify the code in addImgOptMiddleware, because we don't
have to rewrite the URL as aggressively--it's closer to the Fastly
paradigm.

Closes #1808, and some other issues.

test: coverage for addImgOptMiddleware changes